fix: show slack notification toggle for unlinked users too
