+++ infrastructure/hasura/metadata/actions.graphql
@@ -10,6 +10,10 @@ type Mutation {
   invite_user(input: InviteUserInput!): InviteUserOutput
 }
 
+type Query {
+  slack_user(team_id: uuid!): SlackUserOutput
+}
+
 type Mutation {
   upgrade_current_user: UpgradeUserResponse
 }
@@ -60,4 +64,8 @@ type AcceptInvitationsOutput {
 
 type InviteUserOutput {
   success: Boolean!
+}
+
+type SlackUserOutput {
+  slack_user_id: String
 }
