+++ infrastructure/hasura/metadata/actions.yaml
@@ -26,6 +26,15 @@ actions:
       value_from_env: HASURA_ACTIONS_WEBHOOK_AUTHORIZATION_HEADER
   permissions:
   - role: user
+- name: slack_user
+  definition:
+    kind: ""
+    handler: '{{HASURA_ACTIONS_WEBHOOK_URL}}'
+    headers:
+    - name: Authorization
+      value_from_env: HASURA_ACTIONS_WEBHOOK_AUTHORIZATION_HEADER
+  permissions:
+  - role: user
 - name: upgrade_current_user
   definition:
     kind: synchronous
@@ -59,4 +68,5 @@ custom_types:
   - name: FindSlackUserOutput
   - name: AcceptInvitationsOutput
   - name: InviteUserOutput
+  - name: SlackUserOutput
   scalars: []
